xonshrc: drop nix hacks after figuring out poetry2nix
It was always somewhat doomed to fail (eventually) and using the correct
approach (with poetry2nix) for python stuff seems to work just as well
for boostrapping in `cornerboi` + `xonsh`.

The main breaker was that xontrib-vox isn't in nix pkgs and isn't pkged
in general, so there's no way to load it into the system xonsh easily
AND doing `pip install -U` hackery seems to be impossible now due to
the way store's now wrap the pip binary and system/user xonsh.

Summary:
- mask out all the NIX_ var hardcodes instead only leaving in the config
  path map to the config in THIS repo.
- if `xontrib load vox` fails, don't try to load other xontribs and
  instead just comment on the console.
- adjust `pyup()` to instead use `nix develop` in any detected repo dir
  (which means presuming a `flake.nix`) and always generate a hidden
  `.p2nixdev` profile file for faster loading on follow up env usages.
- don't use vox in general for virtualenvs (by masking the old impl) and
  instead always presume we're on nix for now..

Also add some new aliases:
- `esp` for translation to espanol.
- make `trans` always use `-brief`.
- multiline format `nixup` defn.